createPlayerNotificationChannel() fun was moved from App to PlayerServiceImpl class
