+++ app/src/main/java/bes/max/playlistmaker/app/App.kt
@@ -1,18 +1,13 @@
 package bes.max.playlistmaker.app
 
 import android.app.Application
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.asLiveData
-import bes.max.playlistmaker.R
 import bes.max.playlistmaker.data.settings.SettingsDao
 import bes.max.playlistmaker.di.dataModule
 import bes.max.playlistmaker.di.domainModule
 import bes.max.playlistmaker.di.repositoryModule
 import bes.max.playlistmaker.di.viewModelModule
-import bes.max.playlistmaker.domain.player.PlayerService
 import kotlinx.coroutines.launch
 import org.koin.android.ext.android.getKoin
 import org.koin.android.ext.koin.androidContext
@@ -32,8 +27,6 @@ class App : Application() {
         kotlinx.coroutines.MainScope().launch {
             settingsDao.isNightModeActive().asLiveData().observeForever { switchTheme(it) }
         }
-
-        createPlayerNotificationChannel()
     }
 
     private fun switchTheme(darkThemeEnabled: Boolean) {
@@ -43,16 +36,4 @@ class App : Application() {
 
         )
     }
-
-    private fun createPlayerNotificationChannel() {
-        val channel = NotificationChannel(
-            PlayerService.NOTIFICATION_CHANNEL_ID,
-            getString(R.string.player_notification_channel_name),
-            NotificationManager.IMPORTANCE_HIGH
-        )
-        val notificationManager =
-            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(channel)
-    }
-
 }
